PLUGINS-1311: Make CKEditor 5 demo work via tarball installation
CKEditor 5 doesn't allow duplicate packages to exist. When installing all the dependencies from a single `package.json`,
this is usually not an issue, but when symlinking (like we do in these demos via Lerna), the symlinked package has its
own `node_modules`, causing CKEditor 5 dependencies to be duplicated. This makes CKEditor 5 crash.

To circumvent this, I made the `build-dev` command install a tarball of the `@wiris/mathtype-ckeditor5`, and the
`deploy` command revert back to the npm version. This is not an ideal solution as running either command modifies
the `package.json` of the demo. Nonetheless the CKEditor 5 demo works correctly this way.